requirements.txt
Pillow>=10.1.0
imagehash>=4.3.1
reportlab>=4.0.0
pdfplumber>=0.9.0
rapidfuzz>=2.16.0